+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Markbot adheres to [Semantic Versioning](http://semver.org/).
 
 ---
 
-## [Unreleased]
+## [5.3.0] — 2016-01-21
 
 ### Added
 
@@ -17,6 +17,10 @@ Markbot adheres to [Semantic Versioning](http://semver.org/).
 - Removed the `aria-describedat` error ignoring because ARIA 1.1 replaced the attribute with `aria-details`.
 - Removed the `Cache-Control` and `Last-Modified` headers from the web server requests—it was interfering with using the web server as testing & I don’t think (though could use more testing) that it affects the performance metrics.
 - The localhost server now uses PortFinder to find an open port and the port cannot be configured in `package.json` any more.
+
+### Fixed
+
+- Dragging to the Markbot icon in the dock didn’t work if there was already a visible window.
 
 ---